maj com apres ajout d'un com en db
